test: move tests to proper test files
