+++ tests/models/design_electric/conftest.py
@@ -6,68 +6,6 @@ from mocks import MockDAO
 from ramstk.models import RAMSTKDesignElectricRecord
 
 
-@pytest.fixture(scope="function")
-def test_attributes():
-    yield {
-        "revision_id": 1,
-        "hardware_id": 1,
-        "application_id": 0,
-        "area": 0.0,
-        "capacitance": 0.0,
-        "configuration_id": 0,
-        "construction_id": 0,
-        "contact_form_id": 0,
-        "contact_gauge": 0,
-        "contact_rating_id": 0,
-        "current_operating": 0.0,
-        "current_rated": 0.0,
-        "current_ratio": 0.0,
-        "environment_active_id": 0,
-        "environment_dormant_id": 0,
-        "family_id": 0,
-        "feature_size": 0.0,
-        "frequency_operating": 0.0,
-        "insert_id": 0,
-        "insulation_id": 0,
-        "manufacturing_id": 0,
-        "matching_id": 0,
-        "n_active_pins": 0,
-        "n_circuit_planes": 1,
-        "n_cycles": 0,
-        "n_elements": 0,
-        "n_hand_soldered": 0,
-        "n_wave_soldered": 0,
-        "operating_life": 0.0,
-        "overstress": 0,
-        "package_id": 0,
-        "power_operating": 0.0,
-        "power_rated": 0.0,
-        "power_ratio": 0.0,
-        "reason": "",
-        "resistance": 0.0,
-        "specification_id": 0,
-        "technology_id": 0,
-        "temperature_active": 35.0,
-        "temperature_case": 0.0,
-        "temperature_dormant": 25.0,
-        "temperature_hot_spot": 0.0,
-        "temperature_junction": 0.0,
-        "temperature_knee": 25.0,
-        "temperature_rated_max": 0.0,
-        "temperature_rated_min": 0.0,
-        "temperature_rise": 0.0,
-        "theta_jc": 0.0,
-        "type_id": 0,
-        "voltage_ac_operating": 0.0,
-        "voltage_dc_operating": 0.0,
-        "voltage_esd": 0.0,
-        "voltage_rated": 0.0,
-        "voltage_ratio": 0.0,
-        "weight": 0.0,
-        "years_in_production": 1,
-    }
-
-
 @pytest.fixture()
 def mock_program_dao(monkeypatch):
     _design_electric_1 = RAMSTKDesignElectricRecord()
@@ -252,3 +190,76 @@ def mock_program_dao(monkeypatch):
     ]
 
     yield DAO
+
+
+@pytest.fixture(scope="function")
+def test_attributes():
+    yield {
+        "revision_id": 1,
+        "hardware_id": 1,
+        "application_id": 0,
+        "area": 0.0,
+        "capacitance": 0.0,
+        "configuration_id": 0,
+        "construction_id": 0,
+        "contact_form_id": 0,
+        "contact_gauge": 0,
+        "contact_rating_id": 0,
+        "current_operating": 0.0,
+        "current_rated": 0.0,
+        "current_ratio": 0.0,
+        "environment_active_id": 0,
+        "environment_dormant_id": 0,
+        "family_id": 0,
+        "feature_size": 0.0,
+        "frequency_operating": 0.0,
+        "insert_id": 0,
+        "insulation_id": 0,
+        "manufacturing_id": 0,
+        "matching_id": 0,
+        "n_active_pins": 0,
+        "n_circuit_planes": 1,
+        "n_cycles": 0,
+        "n_elements": 0,
+        "n_hand_soldered": 0,
+        "n_wave_soldered": 0,
+        "operating_life": 0.0,
+        "overstress": 0,
+        "package_id": 0,
+        "power_operating": 0.0,
+        "power_rated": 0.0,
+        "power_ratio": 0.0,
+        "reason": "",
+        "resistance": 0.0,
+        "specification_id": 0,
+        "technology_id": 0,
+        "temperature_active": 35.0,
+        "temperature_case": 0.0,
+        "temperature_dormant": 25.0,
+        "temperature_hot_spot": 0.0,
+        "temperature_junction": 0.0,
+        "temperature_knee": 25.0,
+        "temperature_rated_max": 0.0,
+        "temperature_rated_min": 0.0,
+        "temperature_rise": 0.0,
+        "theta_jc": 0.0,
+        "type_id": 0,
+        "voltage_ac_operating": 0.0,
+        "voltage_dc_operating": 0.0,
+        "voltage_esd": 0.0,
+        "voltage_rated": 0.0,
+        "voltage_ratio": 0.0,
+        "weight": 0.0,
+        "years_in_production": 1,
+    }
+
+
+@pytest.fixture(scope="function")
+def test_recordmodel(mock_program_dao):
+    """Get a record model instance for each test function."""
+    dut = mock_program_dao.do_select_all(RAMSTKDesignElectricRecord, _all=False)
+
+    yield dut
+
+    # Delete the device under test.
+    del dut
